==> Python/nb_passers/1_2_Learner_Grouping.ipynb <==
# Learner Grouping
This notebook is used to group learners based on the week they passed the course. Since each course have different grading schedules, we have to manually code them one by one.

import numpy as np
import pandas as pd

## Course FP101x-3T2015
In course FP101x-3T2015, there are 288 problems in total. 3 of them have 0 weight and 1 of them has 2 weights, which means that learners can got 286 points in total. Students who have more than 60% points are passers.

However, in our experiments, we find that there are a small number of learners who have no enough grades but are tagged as "Passers". So I release the passing threshold a little (from 171.6 to 170). Then there are still 8 of 1143 passers who do not have enough grades finally. Since the number of these passers is quite small, we ignore them.

def passweek_detect (df, threshold):
    # list the week, learners' ids are in the indexes 
    weeklist = df.columns.values.tolist()[1:]
    df['total_grade'] = 0
    df['group'] = "NotEnough"
    
    for week in weeklist:
        df['total_grade'] = df['total_grade'] + df.loc[: , week]
        df.ix[(df.total_grade>=threshold) & (df.group=="NotEnough"), 'group'] = week
    
    # return a dataframe with columns: week_1 ... week_n, total_grade, group
    return df

# Read data from csv
learner_grade_byweek = pd.read_csv("FP101x_3T2015_grade_byweek.csv")
print learner_grade_byweek.head(10)
weeklist = learner_grade_byweek.columns.values.tolist()[1:]
print weeklist

# Generate their groups
learner_grade_byweek = passweek_detect(learner_grade_byweek, 170)
print learner_grade_byweek.head(10)

# Output results
learner_grade_byweek = learner_grade_byweek.drop(learner_grade_byweek[learner_grade_byweek.group == "NotEnough"].index)
print learner_grade_byweek.shape
learner_group = learner_grade_byweek[["course_learner_id", "group"]]
print learner_group.head(10)
learner_group.to_csv("FP101x_3T2015_group.csv", index=False)
print learner_group[learner_group.group == "Week_5"].shape
print learner_group[learner_group.group == "Week_6"].shape
print learner_group[learner_group.group == "Week_7"].shape
print learner_group[learner_group.group == "Week_8"].shape

# Generate a new dataframe whose columns are course_learner_id, score, rel_week, group
def trans_grade_byweek(df, weeklist):
    df_trans = pd.DataFrame()
    for week in weeklist:
        data = df.loc[: , ["course_learner_id", week, "group"]]
        data['rel_week'] = week
        data.rename(columns={week: 'score'}, inplace=True)
        df_trans = df_trans.append(data)
    
    df_trans = df_trans.sort_values(['course_learner_id', "rel_week"], ascending=[True, True])
    return df_trans

learner_grade_byweek_alt = pd.read_csv("FP101x_3T2015_grade_byweek_alt.csv")
learner_grade_byweek_alt = learner_grade_byweek_alt.merge(learner_group, on="course_learner_id")
print learner_grade_byweek_alt.head(10)
learner_grade_distribution = trans_grade_byweek(learner_grade_byweek_alt, weeklist)
learner_grade_distribution.to_csv("FP101x_3T2015_grade_distribution.csv", index=False)
print learner_grade_distribution.head(10)

## Course EX101x-3T2015
Course EX101x-3T2015 has a similar grading schedule to FP101x-3T2015. Therefore, we use the same functions defined before.

However, in our experiments, we also find that there are a small number of learners who have no enough grades but are tagged as "Passers". So I release the passing threshold a little (from 105.6 to 105). Then there are still 11 of 1156 passers who do not have enough grades finally. Since the number of these passers is quite small, we ignore them.

# Read data from csv
learner_grade_byweek = pd.read_csv("EX101x_3T2015_grade_byweek.csv")
weeklist = learner_grade_byweek.columns.values.tolist()[1:]

# Generate their groups
learner_grade_byweek = passweek_detect(learner_grade_byweek, 105)

# Output results
learner_grade_byweek = learner_grade_byweek.drop(learner_grade_byweek[learner_grade_byweek.group == "NotEnough"].index)
print learner_grade_byweek.shape
learner_group = learner_grade_byweek[["course_learner_id", "group"]]
print learner_group.head(10)
learner_group.to_csv("EX101x_3T2015_group.csv", index=False)
print learner_group[learner_group.group == "Week_5"].shape
print learner_group[learner_group.group == "Week_6"].shape
print learner_group[learner_group.group == "Week_7"].shape
print learner_group[learner_group.group == "Week_8"].shape

# Transform
learner_grade_byweek_alt = pd.read_csv("EX101x_3T2015_grade_byweek_alt.csv")
learner_grade_byweek_alt = learner_grade_byweek_alt.merge(learner_group, on="course_learner_id")
print learner_grade_byweek_alt.head(10)
learner_grade_distribution = trans_grade_byweek(learner_grade_byweek_alt, weeklist)
learner_grade_distribution.to_csv("EX101x_3T2015_grade_distribution.csv", index=False)
print learner_grade_distribution.head(10)

## Course CTB3365STx-1T2016
In this course, only graded assignments are counted in the final grade (practices are ignored). The are 6 graded assignments from week 2 to week 7. In course grading schedule, final grade will removed 2 lowest scores and only count 4 of them. Therefore, each grade assignment counts for 25% of final grades.

# Read data from csv
learner_grade_byweek_alt = pd.read_csv("CTB3365STx_1T2016_grade_byweek_alt.csv")
print learner_grade_byweek_alt.head(10)
weeklist = learner_grade_byweek_alt.columns.values.tolist()[1:]
print weeklist

learner_grade_byweek_alt["total_grade"] = 0
learner_grade_byweek_alt["group"] = "Week_7"
threshold = 2.4
# week 4
learner_grade_byweek_alt["total_grade"] = (learner_grade_byweek_alt["Week_2"] + 
                                           learner_grade_byweek_alt["Week_3"] + 
                                           learner_grade_byweek_alt["Week_4"])
learner_grade_byweek_alt.ix[(learner_grade_byweek_alt.total_grade>=threshold) & (learner_grade_byweek_alt.group=="Week_7"), 'group'] = "Week_4"
# week 5
learner_grade_byweek_alt["total_grade"] = learner_grade_byweek_alt["total_grade"] + learner_grade_byweek_alt["Week_5"]
learner_grade_byweek_alt.ix[(learner_grade_byweek_alt.total_grade>=threshold) & (learner_grade_byweek_alt.group=="Week_7"), 'group'] = "Week_5"
# week 6
learner_grade_byweek_alt['low_grade'] = learner_grade_byweek_alt[['Week_2', 'Week_3', 'Week_4', 'Week_5', 'Week_6']].min(axis=1)
learner_grade_byweek_alt["total_grade"] = learner_grade_byweek_alt["total_grade"] + learner_grade_byweek_alt["Week_6"] - learner_grade_byweek_alt['low_grade']
learner_grade_byweek_alt.ix[(learner_grade_byweek_alt.total_grade>=threshold) & (learner_grade_byweek_alt.group=="Week_7"), 'group'] = "Week_6"
print learner_grade_byweek_alt.head(10)

learner_group = learner_grade_byweek_alt[["course_learner_id", "group"]]
print learner_group.head(10)
learner_group.to_csv("CTB3365STx_1T2016_group.csv", index=False)
print learner_group[learner_group.group == "Week_4"].shape
print learner_group[learner_group.group == "Week_5"].shape
print learner_group[learner_group.group == "Week_6"].shape
print learner_group[learner_group.group == "Week_7"].shape

learner_grade_distribution = trans_grade_byweek(learner_grade_byweek_alt, weeklist)
learner_grade_distribution.to_csv("CTB3365STx_1T2016_grade_distribution.csv", index=False)
print learner_grade_distribution.head(10)

## Course RI101x-1T2016 (Is this course really suitable in this work?)

In this course, since there is openassessment problems which might be hard to access, we only focus on assignments and assume that all the passers got full grades from openassess problem.

# Read data from csv
learner_grade_byweek_alt = pd.read_csv("RI101x_1T2016_grade_byweek_alt.csv")
print learner_grade_byweek_alt.head(10)
weeklist = learner_grade_byweek_alt.columns.values.tolist()[1:]
print weeklist

def passweek_detect_ri (df, threshold1, threshold2):
    # list the week, learners' ids are in the indexes 
    weeklist = df.columns.values.tolist()[1:]
    df['total_grade'] = 0
    df['group'] = "NotEnough"
    
    lastweek = weeklist[-1]
    for week in weeklist:
        df['total_grade'] = df['total_grade'] + df.loc[: , week]
        df.ix[(df.total_grade>=threshold1) & (df.group=="NotEnough"), 'group'] = week
        if week == lastweek:
            df.ix[(df.total_grade>=threshold2) & (df.group=="NotEnough"), 'group'] = week
        
    # return a dataframe with columns: week_1 ... week_n, total_grade, group
    return df

learner_grade_byweek_alt = passweek_detect_ri(learner_grade_byweek_alt, 4.5*6/7, 3.0*6/7)

print learner_grade_byweek_alt.shape

# Output results
learner_grade_byweek_alt = learner_grade_byweek_alt.drop(learner_grade_byweek_alt[learner_grade_byweek_alt.group == "NotEnough"].index)
print learner_grade_byweek_alt.head(10)
learner_group = learner_grade_byweek_alt[["course_learner_id", "group"]]
print learner_group[learner_group.group == "Week_4"].shape
print learner_group[learner_group.group == "Week_5"].shape
print learner_group[learner_group.group == "Week_6"].shape
print learner_group[learner_group.group == "Week_7"].shape
learner_group.to_csv("RI101x_1T2016_group.csv", index=False)

learner_grade_distribution = trans_grade_byweek(learner_grade_byweek_alt, weeklist)
learner_grade_distribution.to_csv("RI101x_1T2016_grade_distribution.csv", index=False)
print learner_grade_distribution.head(100)
