==> Python/nb_passers/2_1_Video_Processing.ipynb <==
# Video Processing
This notebook is used for generating weekly video statistics for each person.

## Getting learners' information about video interactions from DelftX database
We use following SQL statements getting information about video interactions of all passers. The generated table is named as "CTB3365STx_1T2016_pass_videointeractions.csv"

```sql
SELECT 
	video_interaction.course_learner_id AS course_learner_id,
    video_interaction.video_id AS video_id,
    video_interaction.duration AS duration,
    video_interaction.duration_backward_seek AS duration_backward_seek,
    video_interaction.duration_forward_seek AS duration_forward_seek,
    video_interaction.duration_pause AS duration_pause,
    video_interaction.times_backward_seek AS times_backward_seek,
    video_interaction.times_forward_seek AS times_forward_seek,
    video_interaction.times_pause AS times_pause,
    video_interaction.times_speed_down AS times_speed_down,
    video_interaction.times_speed_up AS times_speed_up,
    video_interaction.start_time AS start_time,
    video_interaction.end_time AS end_time
FROM 
	DelftX2.video_interaction AS video_interaction
	JOIN DelftX2.learner_index AS learner_index
	ON learner_index.course_learner_id = video_interaction.course_learner_id
WHERE # select all the pass user
	video_interaction.course_learner_id IN (
		SELECT 
			course_learner_id 
		FROM 
			DelftX2.course_learner
		WHERE 
			DelftX2.course_learner.certificate_status <> "notpassing")
    AND learner_index.course_id = "course-v1:DelftX+CTB3365STx+1T2016"
```

import pandas as pd
import numpy as np
import os

## Read .csv files 
Read related csv files generated by previous SQL query and the first series of notebook (1_2_Learner_Grouping.ipynb)

# OutputFolder
outputfolder = "video_interaction/"
if not os.path.exists(outputfolder):
    os.makedirs(outputfolder)

# # EX101x relevant csv
# video_interactions = "EX101x_3T2015_pass_videointeractions.csv"
# course_elements = "CourseElements_FPEX.csv"
# learner_group = "EX101x_3T2015_group.csv"
# totol_videolength_byweek = [1238, 1517, 1836, 1176, 1503, 2006, 1873, 1255]

# # FP101x relevant csv
# video_interactions = "FP101x_3T2015_pass_videointeractions.csv"
# course_elements = "CourseElements_FPEX.csv"
# learner_group = "FP101x_3T2015_group.csv"
# totol_videolength_byweek = [2950, 4625, 4105, 2957, 4384, 5326, 1705, 252]

# # CTB3365STx
# video_interactions = "CTB3365STx_1T2016_pass_videointeractions.csv"
# course_elements = "CTB3365STx_1T2016_chapter_split.csv"
# learner_group = "CTB3365STx_1T2016_group.csv"
# totol_videolength_byweek = [789, 8212, 5268, 6358, 5928, 8496, 8360, 457]

# RI101x
video_interactions = "RI101x_1T2016_pass_videointeractions.csv"
course_elements = "RI101x_1T2016_chapter_split.csv"
learner_group = "RI101x_1T2016_group.csv"
totol_videolength_byweek = [3721, 4409, 5622, 5615, 2688, 3280, 4046]

df_video_interactions = pd.read_csv(video_interactions)
# df_video_interactions.rename(columns={'question_id': 'element_id'}, inplace=True)
print df_video_interactions.head(10)

## Merge and aggregate durations of video interactions by week for each learner

# Get course id
course_id = df_video_interactions["course_learner_id"][0].split("_")[0].split(":")[1]
print course_id
# change the format of video id
df_video_interactions["video_id"] = df_video_interactions["video_id"].astype(str)
df_video_interactions["video_id"] = df_video_interactions["video_id"].apply(lambda x: "block-v1:" + course_id + "+type@video+block@" + x)
# print df_video_interactions.head(10)
# change the column name for merging in the next step
df_video_interactions.rename(columns={'video_id': 'element_id'}, inplace=True)
print df_video_interactions.head(10)

# Merge the data
df_course_elements = pd.read_csv(course_elements)
# Set the column name to the same
df_course_elements.rename(columns={'week': 'rel_week'}, inplace=True)
df_course_elements.rename(columns={'relevant_week': 'rel_week'}, inplace=True)
# print df_course_elements.head(10)
df_video_interactions = df_video_interactions.merge(df_course_elements, on="element_id")
# print df_video_interactions.shape
print df_video_interactions.head(10)

# Aggregate the data
df_video_interactions = df_video_interactions[['course_learner_id','rel_week','duration']]
df_video_interactions = df_video_interactions.groupby(['course_learner_id', 'rel_week'], as_index=False).sum()

# replace relevant week values
df_video_interactions['rel_week'] = df_video_interactions['rel_week'].astype(str)
replace_week_name = lambda x: 'Week_' + x
df_video_interactions['rel_week'] = df_video_interactions['rel_week'].apply(replace_week_name)
print df_video_interactions.head(10)

video_interactions_byweek = df_video_interactions.pivot(index='course_learner_id', columns='rel_week', values='duration')
video_interactions_byweek_alt = df_video_interactions.pivot(index='course_learner_id', columns='rel_week', values='duration')
video_interactions_byweek = video_interactions_byweek.fillna(0)
video_interactions_byweek_alt = video_interactions_byweek_alt.fillna(0)
print video_interactions_byweek.head(10)

# Output
course_id_alt = video_interactions.split("_")[0] + "_" + video_interactions.split("_")[1]
outputfile = outputfolder + course_id_alt + "_video_duration_byweek.csv"
video_interactions_byweek.to_csv(outputfile)

# Normalize the value of durations by total length of videos in corresponding weeks
weeklist = video_interactions_byweek_alt.columns.values.tolist()
# print weeklist
for (week, totol_videolength) in zip(weeklist, totol_videolength_byweek):
    print week
    print totol_videolength
    video_interactions_byweek_alt[week] = video_interactions_byweek_alt[week] / totol_videolength
        
print video_interactions_byweek_alt.head(10)

course_id_alt = video_interactions.split("_")[0] + "_" + video_interactions.split("_")[1]
outputfile_alt = outputfolder + course_id_alt + "_video_duration_byweek_alt.csv"
video_interactions_byweek_alt.to_csv(outputfile_alt)

## Calculate average durations and standard deviations of each group of passer 

# Merge with group we calculate in grade processing
df_learner_group = pd.read_csv(learner_group)
print df_learner_group.head(10)
print video_interactions_byweek_alt.head(10)
video_interactions_byweek_alt.reset_index(level=0, inplace=True)
video_interactions_byweek_alt_grouped = video_interactions_byweek_alt.merge(df_learner_group, on="course_learner_id")


# Generate a new dataframe whose columns are course_learner_id, duration, rel_week, group
def trans_video_byweek(df, weeklist):
    df_trans = pd.DataFrame()
    for week in weeklist:
        data = df.loc[: , ["course_learner_id", week, "group"]]
        data['rel_week'] = week
        data.rename(columns={week: 'duration'}, inplace=True)
        df_trans = df_trans.append(data)
    
    df_trans = df_trans.sort_values(['course_learner_id', "rel_week"], ascending=[True, True])
    return df_trans

video_interactions_distribution = trans_video_byweek(video_interactions_byweek_alt_grouped, weeklist)
video_interactions_distribution.reset_index()
print video_interactions_distribution.head(100)
outputfile_video_distribution = outputfolder + course_id_alt + "_video_duration_distribution.csv"
video_interactions_distribution.to_csv(outputfile_video_distribution, index=False)

print video_interactions_distribution.columns.values.tolist()

## Calculate mean and std by group and relevant weeks

video_interactions_avg = video_interactions_distribution[["duration","group","rel_week"]].groupby(['group', 'rel_week'], as_index=False).mean()
video_interactions_avg.rename(columns={'duration': 'avgduration'}, inplace=True)
# print video_interactions_avg.head(10)

video_interactions_std = video_interactions_distribution[["duration","group","rel_week"]].groupby(['group', 'rel_week'], as_index=False).agg(lambda x: np.std(x, ddof=1))
video_interactions_std.rename(columns={'duration': 'stdduration'}, inplace=True)
# print video_interactions_std.head(10)

video_interactions_group_byweek = video_interactions_avg.merge(video_interactions_std, on=["group", "rel_week"])
print video_interactions_group_byweek.head(100)

outputfile_video_group_byweek = outputfolder + course_id_alt + "_video_duration_avg_groupbyweek.csv"
video_interactions_group_byweek.to_csv(outputfile_video_group_byweek, index=False)
