==> Python/nb_passers/3_1_Survey_Answer_Filtering.ipynb <==
# Survey Answer Filtering
This scrips is used for filtering learners' answers in pre-/post-survey of the course.

import pandas as pd
import os, csv

# Read .csv files of learners' answers and anonymization maps

# FP101x-3T2015
course_id = "course-v1:DelftX+FP101x+3T2015"
course_id_alt = "FP101x_3T2015"
anon_file_path = "survey/FP101x_3T2015_anon-ids.csv"
pre_survey_file_path = "survey/FP101x_3T2015_Pre-course_survey.csv"
post_survey_file_path = "survey/FP101x_3T2015_Post-course_survey.csv"

# # EX101x-3T2015
# course_id = "course-v1:DelftX+EX101x+3T2015"
# course_id_alt = "EX101x_3T2015"
# anon_file_path = "survey/EX101x_3T2015_anon-ids.csv"
# pre_survey_file_path = "survey/EX101x_3T2015_Pre-course_survey.csv"
# post_survey_file_path = "survey/EX101x_3T2015_Post-course_survey.csv"

# # CTB3365STx-1T2016
# course_id = "course-v1:DelftX+CTB3365STx+1T2016"
# course_id_alt = "CTB3365STx_1T2016"
# anon_file_path = "survey/CTB3365STx_1T2016_anon-ids.csv"
# pre_survey_file_path = "survey/CTB3365STx_1T2016_Pre-course_survey.csv"
# post_survey_file_path = "survey/CTB3365STx_1T2016_Post-course_survey.csv"

# Output Folder
outputfolderpath = "motivations/"
if not os.path.exists(outputfolderpath):
    os.makedirs(outputfolderpath)

## Read global id and anon id into a map

learner_id_map = {}
anon_file = open(anon_file_path, "r")
id_reader = csv.reader(anon_file)
id_reader.next()
for row in id_reader:
    global_id = row[0].replace("\"","")
    anonymized_id = row[1].replace("\"","")
    learner_id_map[anonymized_id] = global_id
# print learner_id_map

## Read pre-survey from the file

pre_survey_file = open(pre_survey_file_path, "r")
pre_survey_reader = csv.reader(pre_survey_file)

question_id_row = pre_survey_reader.next()
print question_id_row
print len(question_id_row)
question_description_row = pre_survey_reader.next()
print question_description_row
print len(question_description_row)

# Output survey_descriptions
survey_descriptions_file_path = outputfolderpath + course_id_alt + "_pre_survey_description.csv"
with open(survey_descriptions_file_path, 'a') as survey_descriptions_file:
    survey_descriptions_file.write("question_id, question_description" + "\n")
    for i in range(len(list(question_id_row))):
        question_id = course_id + "_pre_" + question_id_row[i].replace("\"","")
        question_description = question_description_row[i].replace("\'", "\\'")
        if question_id != "":
            line = question_id + ", " + question_description
            survey_descriptions_file.write(line + "\n")
        
pre_id_index = list(question_id_row).index("user_id")
# print pre_id_index

# Output survey_responses
survey_responses_file_path = outputfolderpath + course_id_alt + "_pre_survey_reponse.csv"
with open(survey_responses_file_path, 'a') as survey_responses_file:
    survey_responses_file.write("course_learner_id, question_id, reponse" + "\n")
    for row in pre_survey_reader:
        learner_id = row[pre_id_index]
        # print learner_id
        if learner_id in learner_id_map.keys():
            course_learner_id = course_id + "_" + learner_id_map[learner_id]

            for i in range(len(list(question_id_row))):
                question_id = course_id + "_pre_" + question_id_row[i].replace("\"","")
                answer = row[i]
                if answer != "":
                    line = course_learner_id + ", " + question_id + ", " + answer
                    survey_responses_file.write(line + "\n")           

# Close the pre survey file
pre_survey_file.close()

## Read post-survey from the file

post_survey_file = open(post_survey_file_path, "r")
post_survey_reader = csv.reader(post_survey_file)

question_id_row = post_survey_reader.next()
print question_id_row
print len(question_id_row)
question_description_row = post_survey_reader.next()
print question_description_row
print len(question_description_row)

# Output survey_descriptions
survey_descriptions_file_path = outputfolderpath + course_id_alt + "_post_survey_description.csv"
with open(survey_descriptions_file_path, 'a') as survey_descriptions_file:
    survey_descriptions_file.write("question_id, question_description" + "\n")
    for i in range(len(list(question_id_row))):
        question_id = course_id + "_post_" + question_id_row[i].replace("\"","")
        question_description = question_description_row[i].replace("\'", "\\'")
        if question_id != "":
            line = question_id + ", " + question_description
            survey_descriptions_file.write(line + "\n")
        
post_id_index = list(question_id_row).index("user_id")

# Output survey_responses
survey_responses_file_path = outputfolderpath + course_id_alt + "_post_survey_reponse.csv"
with open(survey_responses_file_path, 'a') as survey_responses_file:
    survey_responses_file.write("course_learner_id, question_id, reponse" + "\n")
    for row in post_survey_reader:
        learner_id = row[post_id_index]
        if learner_id in learner_id_map.keys():
            course_learner_id = course_id + "_" + learner_id_map[learner_id]

            for i in range(len(list(question_id_row))):
                question_id = course_id + "_post_" + question_id_row[i].replace("\"","")
                answer = row[i]
                if answer != "":
                    line = course_learner_id + ", " + question_id + ", " + answer
                    survey_responses_file.write(line + "\n") 

# Close the post survey file
post_survey_file.close()
