==> Python/nb_passers/3_2_Motivation_Detection(TODO).ipynb <==
# Generating learners' motivations based on their answers in pre-/post-survey
How to measure learners' motivation?
In our work, we manually clasify answers of pre-/post-survey quations into different types, if a learner has any answer in Type A, we treat the learner belongs to Type A

import os, csv
import pandas as pd

## Read related .csv files into dataframe

# learner_group_file is used to filter passers
# since we only focus on passers in this work

# FP101x-3T2015
learner_group_file_path = ""
# Pre-survey files
pre_survey_description_file_path = ""
pre_survey_response_file_path = ""
# Post-survey files
post_survey_description_file_path = ""
post_survey_response_file_path = ""

# # EX101x-3T2015
# learner_group_file_path = ""
# # Pre-survey files
# pre_survey_description_file_path = ""
# pre_survey_response_file_path = ""
# # Post-survey files
# post_survey_description_file_path = ""
# post_survey_response_file_path = ""

# # CTB3365STx-1T2016
# learner_group_file_path = ""
# # Pre-survey files
# pre_survey_description_file_path = ""
# pre_survey_response_file_path = ""
# # Post-survey files
# post_survey_description_file_path = ""
# post_survey_response_file_path = ""

# Read Data into Dataframe
learner_group = pd.read_csv(learner_group_file_path)
pre_survey_description = pd.read_csv(pre_survey_description_file_path)
pre_survey_response = pd.read_csv(pre_survey_response_file_path)
post_survey_description = pd.read_csv(post_survey_description_file_path)
post_survey_response = pd.read_csv(post_survey_response_file_path)

## Filtering Passers

pre_survey_response_flt = pre_survey_response.merge(learner_group, by="course_learner_id")
post_survey_response_flt = post_survey_response.merge(learner_group, by="course_learner_id")

## Filtering Related Answers

# Filter the survey description, since only a few questions are related to our analysis 
question_list = [""]
pre_survey_description_flt = ??
post_survey_description_flt = ??

# Merge the result of filtered survey questions and answers
pre_survey_response_flt = pre_survey_response_flt.merge(pre_survey_description_flt)
post_survey_response_flt = post_survey_response_flt.merge(post_survey_description_flt)
