==> Python/nb_passers/3_3_Passer_Info_Extraction.ipynb <==
# Extract the demographics and paying info of Passers

The main purpose of this script is to generate a file for passers' features.

The format of output csv file is ```course_learner_id, gender, age, edu, national, payment```

NULL or na will be filled as ```Unknown```

Some barcharts or piecharts are used to explore how the data looks like.

``` sql
SELECT 
	learner.course_learner_id AS course_learner_id,
    learner.enrollment_mode AS enrollment_mode,
    demographic.gender AS gender,
    demographic.year_of_birth AS year_of_birth,
    demographic.level_of_education AS level_of_education,
    demographic.country AS country
FROM 
	DelftX.course_learner AS learner
    JOIN DelftX.learner_demographic AS demographic
	ON learner.course_learner_id = demographic.course_learner_id
WHERE 
	learner.certificate_status <> "notpassing")
```

import pandas as pd

# Read data
